refactor: change file name to users
